Streams

+++ OS/include/types.h
@@ -1,7 +1,7 @@
 #ifndef TYPES_H
 #define TYPES_H
 
-typedef unsigned long           uint64_t;
+typedef unsigned long long      uint64_t;
 typedef unsigned int            uint32_t;
 typedef unsigned short          uint16_t;
 typedef unsigned char           uint8_t;
@@ -9,7 +9,7 @@ typedef long long               int64_t;
 typedef int                     int32_t;
 typedef short                   int16_t;
 typedef char                    int8_t;
-typedef uint64_t                size_t;
+typedef unsigned long           size_t;
 
 typedef uint64_t*               uintptr_t;
 typedef uint8_t                 byte_t;
